chore: rename utils.py to core.py
